fix: debugging the sequential bug for x/y order for the wind-interpolation model (adding this function to the public space)
