+++ src/geodata/model/pvlib/_base.py
@@ -34,34 +34,11 @@ from pvlib.location import Location
 from pvlib.modelchain import ModelChain
 from timezonefinder import TimezoneFinder
 
-import numpy as np
-
-from .._base import BaseModel, _should_use_parallel_reading
+from .._base import BaseModel, _normalize_slice_for_sel, _should_use_parallel_reading
 from geodata.logging import logger
 from .calculations import calculate_pvlib_solarposition, calculate_ghi, calculate_relative_humidity, calculate_precipitable_water, convert_kelvin_to_celsius
 from tqdm.auto import tqdm
 
-
-def _normalize_slice_for_sel(coord: xr.DataArray, s: slice) -> slice:
-    """Return a slice that selects the intended coordinate range regardless of bound order or dim direction.
-
-    - xarray's .sel(dim=slice(a, b)) returns empty when the dimension is descending (e.g. ERA5
-      latitude) or when the user passes slice(high, low) on an ascending dimension (e.g. slice(125, 114.5)).
-    - This helper always interprets the slice as the logical range [min(start, stop), max(start, stop)]
-      and returns slice bounds in the order required by .sel() for the coordinate's direction.
-    """
-    if not isinstance(s, slice) or s.step not in (None, 1):
-        return s
-    if s.start is None or s.stop is None:
-        return s
-    lo, hi = min(s.start, s.stop), max(s.start, s.stop)
-    vals = np.asarray(coord.values).ravel()
-    if len(vals) < 2:
-        return slice(lo, hi)
-    descending = np.all(np.diff(vals) <= 0)
-    if descending:
-        return slice(hi, lo)
-    return slice(lo, hi)
 
 class ModelChainConfig:
     """
